requirements.txt
# Requirements for Vehicle Painter
# tkinter is included with Python standard library, no need to install separately
Pillow>=10.0.0  # For image handling and tileset support

